Updating docs, addressing feedback
