+++ sdk/appcenter-analytics/src/main/java/com/microsoft/appcenter/analytics/PropertyConfigurator.java
@@ -1,51 +1,57 @@
 package com.microsoft.appcenter.analytics;
 
 import android.support.annotation.NonNull;
-import android.support.annotation.VisibleForTesting;
-import android.util.Property;
 
-import com.microsoft.appcenter.analytics.ingestion.models.one.CommonSchemaEventLog;
-import com.microsoft.appcenter.analytics.ingestion.models.one.json.CommonSchemaEventLogFactory;
 import com.microsoft.appcenter.channel.AbstractChannelListener;
 import com.microsoft.appcenter.channel.Channel;
-import com.microsoft.appcenter.channel.OneCollectorChannelListener;
-import com.microsoft.appcenter.ingestion.models.CustomPropertiesLog;
 import com.microsoft.appcenter.ingestion.models.Log;
-import com.microsoft.appcenter.ingestion.models.StartServiceLog;
-import com.microsoft.appcenter.ingestion.models.json.CustomPropertiesLogFactory;
-import com.microsoft.appcenter.ingestion.models.json.DefaultLogSerializer;
-import com.microsoft.appcenter.ingestion.models.json.LogSerializer;
-import com.microsoft.appcenter.ingestion.models.json.StartServiceLogFactory;
 import com.microsoft.appcenter.ingestion.models.one.AppExtension;
 import com.microsoft.appcenter.ingestion.models.one.CommonSchemaLog;
-import com.microsoft.appcenter.ingestion.models.one.SdkExtension;
-import com.microsoft.appcenter.utils.AppCenterLog;
-import com.microsoft.appcenter.utils.UUIDUtils;
-
-import java.util.Collection;
-
-import static com.microsoft.appcenter.utils.AppCenterLog.LOG_TAG;
 
 /**
  * Allow overriding Part A properties.
  */
 public class PropertyConfigurator extends AbstractChannelListener {
 
+    /**
+     * App name to override common schema part A 'app.name'.
+     */
     private String mAppName;
 
+    /**
+     * App name to override common schema part A 'app.ver'.
+     */
     private String mAppVersion;
 
+    /**
+     * App name to override common schema part A 'app.locale'.
+     */
     private String mAppLocale;
 
+    /**
+     * The transmission target which this configurator belongs to.
+     */
     private AnalyticsTransmissionTarget mTransmissionTarget;
 
+    /**
+     * Create a new property configurator.
+     *
+     * @param channel            The channel for this listener.
+     * @param transmissionTarget The tranmission target of the configurator.
+     */
     PropertyConfigurator(Channel channel, AnalyticsTransmissionTarget transmissionTarget) {
         mTransmissionTarget = transmissionTarget;
         if (channel != null) {
             channel.addListener(this);
         }
     }
-    
+
+    /**
+     * Override or inherit common schema properties while preparing log.
+     *
+     * @param log       A log.
+     * @param groupName The group name.
+     */
     @Override
     public void onPreparingLog(@NonNull Log log, @NonNull String groupName) {
         if (log instanceof CommonSchemaLog && mTransmissionTarget.isEnabled()) {
@@ -55,14 +61,12 @@ public class PropertyConfigurator extends AbstractChannelListener {
             if (mAppName != null) {
                 app.setId(mAppName);
             } else {
-                AnalyticsTransmissionTarget parent = mTransmissionTarget.mParentTarget;
-                while (parent != null) {
-                    String parentAppName = parent.getPropertyConfigurator().getAppName();
+                for (AnalyticsTransmissionTarget target = mTransmissionTarget.mParentTarget; target != null; target = target.mParentTarget) {
+                    String parentAppName = target.getPropertyConfigurator().getAppName();
                     if (parentAppName != null) {
                         app.setId(parentAppName);
                         break;
                     }
-                    parent = parent.mParentTarget;
                 }
             }
 
@@ -70,14 +74,12 @@ public class PropertyConfigurator extends AbstractChannelListener {
             if (mAppVersion != null) {
                 app.setVer(mAppVersion);
             } else {
-                AnalyticsTransmissionTarget parent = mTransmissionTarget.mParentTarget;
-                while (parent != null) {
-                    String parentAppVersion = parent.getPropertyConfigurator().getAppVersion();
+                for (AnalyticsTransmissionTarget target = mTransmissionTarget.mParentTarget; target != null; target = target.mParentTarget) {
+                    String parentAppVersion = target.getPropertyConfigurator().getAppVersion();
                     if (parentAppVersion != null) {
                         app.setVer(parentAppVersion);
                         break;
                     }
-                    parent = parent.mParentTarget;
                 }
             }
 
@@ -85,55 +87,68 @@ public class PropertyConfigurator extends AbstractChannelListener {
             if (mAppLocale != null) {
                 app.setLocale(mAppLocale);
             } else {
-                AnalyticsTransmissionTarget parent = mTransmissionTarget.mParentTarget;
-                while (parent != null) {
-                    String parentAppLocale = parent.getPropertyConfigurator().getAppLocale();
+                for (AnalyticsTransmissionTarget target = mTransmissionTarget.mParentTarget; target != null; target = target.mParentTarget) {
+                    String parentAppLocale = target.getPropertyConfigurator().getAppLocale();
                     if (parentAppLocale != null) {
                         app.setLocale(parentAppLocale);
                         break;
                     }
-                    parent = parent.mParentTarget;
                 }
             }
         }
     }
 
     /**
+     * Get app name. Used for checking parents for property inheritance.
+     *
+     * @return App name
+     */
+    private String getAppName() {
+        return mAppName;
+    }
+
+    /**
      * Override common schema Part A property App.Name.
      *
-     * @param appName   App name
+     * @param appName App name
      */
     public void setAppName(String appName) {
         mAppName = appName;
     }
 
-    protected String getAppName() {
-        return mAppName;
+    /**
+     * Get app version. Used for checking parents for property inheritance.
+     *
+     * @return App version
+     */
+    private String getAppVersion() {
+        return mAppVersion;
     }
 
     /**
      * Override common schema Part A property App.Version.
      *
-     * @param appVersion    App version
+     * @param appVersion App version
      */
     public void setAppVersion(String appVersion) {
         mAppVersion = appVersion;
     }
 
-    protected String getAppVersion() {
-        return mAppVersion;
+    /**
+     * Get app locale. Used for checking parents for property inheritance.
+     *
+     * @return App locale
+     */
+    private String getAppLocale() {
+        return mAppLocale;
     }
 
     /**
      * Override common schema Part A property App.Locale.
      *
-     * @param appLocale     App Locale
+     * @param appLocale App Locale
      */
     public void setAppLocale(String appLocale) {
         mAppLocale = appLocale;
-    }
-
-    protected String getAppLocale() {
-        return mAppLocale;
     }
 }
