+++ sdk/appcenter-analytics/src/test/java/com/microsoft/appcenter/analytics/AnalyticsTransmissionTargetTest.java
@@ -3,15 +3,9 @@ package com.microsoft.appcenter.analytics;
 import android.content.Context;
 
 import com.microsoft.appcenter.analytics.ingestion.models.EventLog;
-import com.microsoft.appcenter.analytics.ingestion.models.one.CommonSchemaEventLog;
 import com.microsoft.appcenter.channel.Channel;
 import com.microsoft.appcenter.ingestion.models.Log;
-import com.microsoft.appcenter.ingestion.models.one.AppExtension;
-import com.microsoft.appcenter.ingestion.models.one.CommonSchemaLog;
-import com.microsoft.appcenter.ingestion.models.one.Extensions;
 
-import org.hamcrest.core.CombinableMatcher;
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
@@ -24,7 +18,6 @@ import java.util.Map;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertSame;
@@ -166,7 +159,6 @@ public class AnalyticsTransmissionTargetTest extends AbstractAnalyticsTest {
         /* Another child transmission target with the same token should be the same instance. */
         assertSame(childTarget, target.getTransmissionTarget("token3"));
     }
-
 
     @Test
     public void setEnabled() {
@@ -442,99 +434,5 @@ public class AnalyticsTransmissionTargetTest extends AbstractAnalyticsTest {
         expectedProperties.put("f", "6666");
         expectedProperties.put("g", "7777");
         assertEquals(expectedProperties, log.getProperties());
-    }
-
-    /* Property Configurator tests */
-    @Test
-    public void setCommonSchemaProperties() {
-        CommonSchemaLog log = new CommonSchemaEventLog();
-        log.setExt(new Extensions());
-        log.getExt().setApp(new AppExtension());
-
-        /* Get property configurator and set properties. */
-        PropertyConfigurator pc = Analytics.getTransmissionTarget("test").getPropertyConfigurator();
-        pc.setAppVersion("appVersion");
-        pc.setAppName("appName");
-        pc.setAppLocale("appLocale");
-        pc.onPreparingLog(log, "groupName");
-
-        /* Assert properties set on common schema. */
-        assertEquals("appVersion", log.getExt().getApp().getVer());
-        assertEquals("appName", log.getExt().getApp().getId());
-        assertEquals("appLocale", log.getExt().getApp().getLocale());
-    }
-
-    @Test
-    public void commonSchemaPropertiesNotSetWhenDisabled() {
-        CommonSchemaLog log = new CommonSchemaEventLog();
-        log.setExt(new Extensions());
-        log.getExt().setApp(new AppExtension());
-
-        /* Get target, disable it, and set properties. */
-        AnalyticsTransmissionTarget target = Analytics.getTransmissionTarget("test");
-        target.setEnabledAsync(false);
-        target.getPropertyConfigurator().setAppVersion("appVersion");
-        target.getPropertyConfigurator().setAppName("appName");
-        target.getPropertyConfigurator().setAppLocale("appLocale");
-        target.getPropertyConfigurator().onPreparingLog(log, "groupName");
-
-        /* Assert properties are null. */
-        assertNull(log.getExt().getApp().getVer());
-        assertNull(log.getExt().getApp().getId());
-        assertNull(log.getExt().getApp().getLocale());
-    }
-
-    @Test
-    public void inheritCommonSchemaPropertiesFromGrandparent() {
-        CommonSchemaLog log = new CommonSchemaEventLog();
-        log.setExt(new Extensions());
-        log.getExt().setApp(new AppExtension());
-
-        /* Set properties on parent to override unset properties on child */
-        AnalyticsTransmissionTarget grandparent = Analytics.getTransmissionTarget("grandparent");
-        grandparent.getPropertyConfigurator().setAppVersion("appVersion");
-        grandparent.getPropertyConfigurator().setAppName("appName");
-        grandparent.getPropertyConfigurator().setAppLocale("appLocale");
-
-        AnalyticsTransmissionTarget parent = grandparent.getTransmissionTarget("parent");
-        AnalyticsTransmissionTarget child = parent.getTransmissionTarget("child");
-        child.getPropertyConfigurator().onPreparingLog(log, "groupName");
-
-        /* Assert properties set on common schema. */
-        assertEquals("appVersion", log.getExt().getApp().getVer());
-        assertEquals("appName", log.getExt().getApp().getId());
-        assertEquals("appLocale", log.getExt().getApp().getLocale());
-    }
-
-    @Test
-    public void grandparentsHaveNoPropertiesSet() {
-        CommonSchemaLog log = new CommonSchemaEventLog();
-        log.setExt(new Extensions());
-        log.getExt().setApp(new AppExtension());
-
-        /* Set up empty chain of parents. */
-        AnalyticsTransmissionTarget grandparent = Analytics.getTransmissionTarget("grandparent");
-        AnalyticsTransmissionTarget parent = grandparent.getTransmissionTarget("parent");
-        AnalyticsTransmissionTarget child = parent.getTransmissionTarget("child");
-        child.getPropertyConfigurator().onPreparingLog(log, "groupName");
-
-        /* Assert properties set on common schema. */
-        assertNull(log.getExt().getApp().getVer());
-        assertNull(log.getExt().getApp().getId());
-        assertNull(log.getExt().getApp().getLocale());
-    }
-
-    @Test
-    public void appCenterLogDoesNotOverride() {
-        Log log = new EventLog();
-
-        /* Get property configurator and set properties. */
-        PropertyConfigurator pc = Analytics.getTransmissionTarget("test").getPropertyConfigurator();
-        pc.setAppVersion("appVersion");
-        pc.setAppName("appName");
-        pc.setAppLocale("appLocale");
-        pc.onPreparingLog(log, "groupName");
-
-        verify(mChannel, never()).enqueue(isA(EventLog.class), anyString());
     }
 }
